Run Go migration functions
Sometimes it's useful to do this, if you need to do some complex
operations on data. Previously this was run after all the regular
migrations in GoatCounter, but this isn't ideal as it's not in order. So
if regular migration B depends on Go migration A you'll run in to
trouble.

This is a bit ugly; but it works for now.