Exec() no longer returns sql.Result
Use InsertID() and NumRows() instead. 99% of the time you don't care
about it, and writing "_, err := Exec()" all the time is just annoying.

This also makes it easier for people to use something like
must(Exec(...)), where must() panics or t.Fatal()s on errors. Actually,
we may want to add something like that for tests.