Add zdb.SchemaTemplate(), zdb.Driver()
With SchemaTemplate it's a bit easier to write compatible SQL schemas. I
was go generate-ing them before, which is a bit inconvenient.

With Driver() it's easier to get the SQL driver. The DriverName() isn't
really enough as people can register custom drivers and the like. This
removes the SQLite() and PgSQL() functions.