Import sqlx in internal/sqlx
There are a bunch of bugs and issues with sqlx, and the project is not
very actively maintained. I'd send PRs, but the hopes of getting them
merged are slim, and for some issues there are already open (unmerged)
PRs.

So just stick it in internal for now, so we can address some issues.