Move to "driver" layout; change connection string to "dialect+connect"
The "URL-style" caused confusion because if PostgreSQL's URL-style
connection string, which has been a long-standing issue. See e.g. #2.

Changing the syntax isn't great because it's incompatible, but better to
change it now that only GoatCounter is using this package (AFAIK
anyway).

Also start moving database-engine specific stuff to drivers, so we don't
need to depend on specific SQL drivers. Still depend on them for tests,
not super-great but okay.