Close pgxpool.Pool
Just closing the sql.DB isn't enough, as that won't close all the
connections made from pgxpool.Pool. Need to explicitly close both.

This should really be done on db.Close() for everything, but not so easy
to wrap that as sql.DB is a struct rather than interface. Would probably
be best to send a patch to pgx to add option or something. Need to look
into it properly.

However, in "real applications" it's usually not a big deal as
connections as only closed on application exit (that is: basically
never). So for now, it's an okay hack to just do this for tests.