Create database by connecting to "postgres" database
Rather than using "createdb" from $PATH. This is more convenient as it
won't require some local PostgreSQL installation.

We can only do this from the pgx river, as pq doesn't really offer the
necessary connection string parse bits.

Also remove the "exists" return; this was/is used to decide whether to
run the schema.sql or not, but with the hasTables() check that's kind of
superfluous.