Move drivers to zgo.at/zdb-drivers
Already started work on this some years ago, but left unfinished because
getting the tests to work is not so straight-forward.

I don't want zgo.at/zdb to depend on anything, but to actually test
something meaningful you need to import a driver. Adding to the problems
is that drivers themselves depend on zgo.at/zdb.

To solve this, move most tests to a new "test" package with its own
go.mod. It's not ideal, but it works.